Move quiz generation logic to consumer and rename files
