+++ server/redpanda/consumers/QuizResponseConsumer.ts
